[Stack monitoring] Remove getAngularInjector and duplicated angular components (#115593)
* remove getAngularInjector and old angular components

* Remove suffix from CcrShardReact component

* Remove suffix from ElasticsearchOverviewReact component

* Remove suffix from indexReact component

* Remove suffix from NodeReact component

* Remove suffix from ShardActivityReact

* Remove suffix from ShardAllocationReact component and its childs

* Fix import

* fix translations

# Conflicts:
#	x-pack/plugins/monitoring/public/components/elasticsearch/node/node.js
