+++ x-pack/plugins/monitoring/public/application/pages/elasticsearch/ccr_shard_page.tsx
@@ -13,7 +13,7 @@ import { PageTemplate } from '../page_template';
 import { useKibana } from '../../../../../../../src/plugins/kibana_react/public';
 import { GlobalStateContext } from '../../contexts/global_state_context';
 // @ts-ignore
-import { CcrShardReact } from '../../../components/elasticsearch/ccr_shard';
+import { CcrShard } from '../../../components/elasticsearch/ccr_shard';
 import { ComponentProps } from '../../route_init';
 import { SetupModeRenderer, SetupModeProps } from '../../../components/renderers/setup_mode';
 import { SetupModeContext } from '../../../components/setup_mode/setup_mode_context';
@@ -113,7 +113,7 @@ export const ElasticsearchCcrShardPage: React.FC<ComponentProps> = ({ clusters }
         render={({ flyoutComponent, bottomBarComponent }: SetupModeProps) => (
           <SetupModeContext.Provider value={{ setupModeSupported: true }}>
             {flyoutComponent}
-            <CcrShardReact {...data} alerts={alerts} />
+            <CcrShard {...data} alerts={alerts} />
             {bottomBarComponent}
           </SetupModeContext.Provider>
         )}
