+++ x-pack/plugins/monitoring/public/application/pages/elasticsearch/index_page.tsx
@@ -11,7 +11,7 @@ import { find } from 'lodash';
 import { useKibana } from '../../../../../../../src/plugins/kibana_react/public';
 import { GlobalStateContext } from '../../contexts/global_state_context';
 // @ts-ignore
-import { IndexReact } from '../../../components/elasticsearch/index/index_react';
+import { Index } from '../../../components/elasticsearch/index/index';
 import { ComponentProps } from '../../route_init';
 import { SetupModeRenderer, SetupModeProps } from '../../../components/renderers/setup_mode';
 import { SetupModeContext } from '../../../components/setup_mode/setup_mode_context';
@@ -118,7 +118,7 @@ export const ElasticsearchIndexPage: React.FC<ComponentProps> = ({ clusters }) =
         render={({ setupMode, flyoutComponent, bottomBarComponent }: SetupModeProps) => (
           <SetupModeContext.Provider value={{ setupModeSupported: true }}>
             {flyoutComponent}
-            <IndexReact
+            <Index
               setupMode={setupMode}
               labels={indexLabel}
               alerts={alerts}
