+++ x-pack/plugins/monitoring/public/application/pages/elasticsearch/node_page.tsx
@@ -11,7 +11,7 @@ import { i18n } from '@kbn/i18n';
 import { ItemTemplate } from './item_template';
 import { useKibana } from '../../../../../../../src/plugins/kibana_react/public';
 import { GlobalStateContext } from '../../contexts/global_state_context';
-import { NodeReact } from '../../../components/elasticsearch';
+import { Node } from '../../../components/elasticsearch';
 import { ComponentProps } from '../../route_init';
 import { SetupModeRenderer, SetupModeProps } from '../../../components/renderers/setup_mode';
 import { SetupModeContext } from '../../../components/setup_mode/setup_mode_context';
@@ -143,7 +143,7 @@ export const ElasticsearchNodePage: React.FC<ComponentProps> = ({ clusters }) =>
         render={({ setupMode, flyoutComponent, bottomBarComponent }: SetupModeProps) => (
           <SetupModeContext.Provider value={{ setupModeSupported: true }}>
             {flyoutComponent}
-            <NodeReact
+            <Node
               alerts={alerts}
               labels={labels.node}
               nodeId={node}
