+++ x-pack/plugins/monitoring/public/application/pages/elasticsearch/overview.tsx
@@ -11,7 +11,7 @@ import { ElasticsearchTemplate } from './elasticsearch_template';
 import { useKibana } from '../../../../../../../src/plugins/kibana_react/public';
 import { GlobalStateContext } from '../../contexts/global_state_context';
 // @ts-ignore
-import { ElasticsearchOverviewReact } from '../../../components/elasticsearch';
+import { ElasticsearchOverview } from '../../../components/elasticsearch';
 import { ComponentProps } from '../../route_init';
 import { useCharts } from '../../hooks/use_charts';
 import { BreadcrumbContainer } from '../../hooks/use_breadcrumbs';
@@ -79,7 +79,7 @@ export const ElasticsearchOverviewPage: React.FC<ComponentProps> = ({ clusters }
     const shardActivityData = shardActivity && filterShardActivityData(shardActivity); // no filter on data = null
 
     return (
-      <ElasticsearchOverviewReact
+      <ElasticsearchOverview
         clusterStatus={clusterStatus}
         metrics={metrics}
         logs={logs}
