+++ x-pack/plugins/monitoring/public/components/elasticsearch/ccr_shard/index.js
@@ -6,4 +6,3 @@
  */
 
 export { CcrShard } from './ccr_shard';
-export { CcrShardReact } from './ccr_shard_react';
