+++ x-pack/plugins/monitoring/public/components/elasticsearch/node/index.ts
@@ -7,4 +7,3 @@
 
 export { NodeStatusIcon } from './status_icon';
 export { Node } from './node';
-export { NodeReact } from './node_react';
