+++ x-pack/plugins/monitoring/public/components/elasticsearch/node/node.js
@@ -15,22 +15,14 @@ import {
   EuiFlexItem,
   EuiPanel,
 } from '@elastic/eui';
+import { FormattedMessage } from '@kbn/i18n/react';
 import { NodeDetailStatus } from '../node_detail_status';
-import { Logs } from '../../logs/';
+import { Logs } from '../../logs';
 import { MonitoringTimeseriesContainer } from '../../chart';
-import { ShardAllocation } from '../shard_allocation/shard_allocation';
 import { AlertsCallout } from '../../../alerts/callout';
+import { ShardAllocation } from '../shard_allocation';
 
-export const Node = ({
-  nodeSummary,
-  metrics,
-  logs,
-  alerts,
-  nodeId,
-  clusterUuid,
-  scope,
-  ...props
-}) => {
+export const Node = ({ nodeSummary, metrics, logs, alerts, nodeId, clusterUuid, ...props }) => {
   /*
   // This isn't doing anything due to a possible bug.  https://github.com/elastic/kibana/issues/106309
   if (alerts) {
@@ -82,7 +74,7 @@ export const Node = ({
         </EuiPanel>
         <EuiSpacer size="m" />
         <EuiPanel>
-          <ShardAllocation scope={scope} />
+          <ShardAllocation {...props} />
         </EuiPanel>
       </EuiPageBody>
     </EuiPage>
