--- x-pack/plugins/monitoring/public/components/elasticsearch/node/node_react.d.ts
@@ -1,19 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { FunctionComponent } from 'react';
-
-export const NodeReact: FunctionComponent<NodeReactProps>;
-export interface NodeReactProps {
-  nodeSummary: unknown;
-  metrics: unknown;
-  logs: unknown;
-  alerts: unknown;
-  nodeId: unknown;
-  clusterUuid: unknown;
-  [key: string]: any;
-}
