+++ x-pack/plugins/monitoring/public/components/elasticsearch/overview/index.ts
@@ -5,6 +5,5 @@
  * 2.0.
  */
 
-export { ElasticsearchOverview } from './overview';
 // @ts-ignore
-export { ElasticsearchOverviewReact } from './overview_react';
+export { ElasticsearchOverview } from './overview';
