+++ x-pack/plugins/monitoring/public/components/elasticsearch/shard_activity/index.js
@@ -6,4 +6,3 @@
  */
 
 export { ShardActivity } from './shard_activity';
-export { ShardActivityReact } from './shard_activity_react';
