+++ x-pack/plugins/monitoring/public/components/elasticsearch/shard_allocation/components/cluster_view.js
@@ -8,67 +8,22 @@
 import React from 'react';
 import { TableHead } from './table_head';
 import { TableBody } from './table_body';
-import { i18n } from '@kbn/i18n';
 
-export class ClusterView extends React.Component {
-  static displayName = i18n.translate(
-    'xpack.monitoring.elasticsearch.shardAllocation.clusterViewDisplayName',
-    {
-      defaultMessage: 'ClusterView',
-    }
+export const ClusterView = (props) => {
+  return (
+    <table cellPadding="0" cellSpacing="0" className="table">
+      <TableHead
+        labels={props.labels}
+        toggleShowSystemIndices={props.toggleShowSystemIndices}
+        showSystemIndices={props.showSystemIndices}
+      />
+      <TableBody
+        filter={props.filter}
+        totalCount={props.totalCount}
+        rows={props.nodesByIndices}
+        cols={props.labels.length}
+        shardStats={props.shardStats}
+      />
+    </table>
   );
-
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      labels: props.scope.labels || [],
-      showing: props.scope.showing || [],
-      shardStats: props.scope.pageData.shardStats,
-      showSystemIndices: props.showSystemIndices,
-      toggleShowSystemIndices: props.toggleShowSystemIndices,
-    };
-  }
-
-  setShowing = (data) => {
-    if (data) {
-      this.setState({ showing: data });
-    }
-  };
-
-  setShardStats = (stats) => {
-    this.setState({ shardStats: stats });
-  };
-
-  UNSAFE_componentWillMount() {
-    this.props.scope.$watch('showing', this.setShowing);
-    this.props.scope.$watch(() => this.props.scope.pageData.shardStats, this.setShardStats);
-  }
-
-  hasUnassigned = () => {
-    return (
-      this.state.showing.length &&
-      this.state.showing[0].unassigned &&
-      this.state.showing[0].unassigned.length
-    );
-  };
-
-  render() {
-    return (
-      <table cellPadding="0" cellSpacing="0" className="table">
-        <TableHead
-          hasUnassigned={this.hasUnassigned()}
-          scope={this.props.scope}
-          toggleShowSystemIndices={this.state.toggleShowSystemIndices}
-        />
-        <TableBody
-          filter={this.props.scope.filter}
-          totalCount={this.props.scope.totalCount}
-          rows={this.state.showing}
-          cols={this.state.labels.length}
-          shardStats={this.state.shardStats}
-        />
-      </table>
-    );
-  }
-}
+};
