+++ x-pack/plugins/monitoring/public/components/logs/logs.js
@@ -16,18 +16,8 @@ import { FormattedMessage } from '@kbn/i18n/react';
 import { Reason } from './reason';
 
 const getFormattedDateTimeLocal = (timestamp) => {
-  try {
-    const injector = Legacy.shims.getAngularInjector();
-    const timezone = injector.get('config').get('dateFormat:tz');
-    return formatDateTimeLocal(timestamp, timezone);
-  } catch (error) {
-    if (error.message === 'Angular has been removed.') {
-      const timezone = Legacy.shims.uiSettings?.get('dateFormat:tz');
-      return formatDateTimeLocal(timestamp, timezone);
-    } else {
-      throw error;
-    }
-  }
+  const timezone = Legacy.shims.uiSettings?.get('dateFormat:tz');
+  return formatDateTimeLocal(timestamp, timezone);
 };
 
 const columnTimestampTitle = i18n.translate('xpack.monitoring.logs.listing.timestampTitle', {
