+++ x-pack/plugins/translations/translations/ja-JP.json
@@ -18201,7 +18201,6 @@
     "xpack.monitoring.elasticsearch.shardActivity.totalTimeTooltip": "開始：{startTime}",
     "xpack.monitoring.elasticsearch.shardActivity.unknownTargetAddressContent": "不明",
     "xpack.monitoring.elasticsearch.shardActivityTitle": "シャードアクティビティ",
-    "xpack.monitoring.elasticsearch.shardAllocation.clusterViewDisplayName": "ClusterView",
     "xpack.monitoring.elasticsearch.shardAllocation.decorateShards.relocatingFromTextMessage": "{nodeName} から移動しています",
     "xpack.monitoring.elasticsearch.shardAllocation.decorateShards.relocatingToTextMessage": "{nodeName} に移動しています",
     "xpack.monitoring.elasticsearch.shardAllocation.initializingLabel": "初期化中",
